ml-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
numpy==1.26.2
pandas==2.1.4
scikit-learn==1.3.2
xgboost==2.0.3
transformers==4.36.2
torch==2.1.2
spacy==3.7.2
shap==0.44.0
prophet==1.1.5
joblib==1.3.2
python-dotenv==1.0.0
langdetect==1.0.9
redis==5.0.1
